fix: new contact check for email match fixed
