+++ Backend/src/main/java/mygoogleserviceapi/contacts/service/ContactListServiceImpl.java
@@ -1,17 +1,9 @@
 package mygoogleserviceapi.contacts.service;
 
 import lombok.RequiredArgsConstructor;
-import mygoogleserviceapi.contacts.dto.ContactDTO;
-import mygoogleserviceapi.contacts.dto.ContactEmailAddressDTO;
-import mygoogleserviceapi.contacts.dto.ContactPhoneNumberDTO;
-import mygoogleserviceapi.contacts.enumeration.ContactTypeEnum;
 import mygoogleserviceapi.contacts.model.Contact;
-import mygoogleserviceapi.contacts.model.ContactEmailAddress;
 import mygoogleserviceapi.contacts.model.ContactList;
-import mygoogleserviceapi.contacts.model.ContactPhoneNumber;
-import mygoogleserviceapi.contacts.repository.ContactEmailAddressRepository;
 import mygoogleserviceapi.contacts.repository.ContactListRepository;
-import mygoogleserviceapi.contacts.repository.ContactPhoneNumberRepository;
 import mygoogleserviceapi.contacts.repository.ContactRepository;
 import mygoogleserviceapi.contacts.service.interfaces.ContactListService;
 import mygoogleserviceapi.contacts.service.interfaces.ContactPictureStorageService;
@@ -22,10 +14,7 @@ import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -37,8 +26,6 @@ public class ContactListServiceImpl implements ContactListService {
     private final FileStorageService fileStorageService;
     private final ContactListRepository contactListRepository;
     private final ContactPictureStorageService contactPictureStorageService;
-    private final ContactPhoneNumberRepository contactPhoneNumberRepository;
-    private final ContactEmailAddressRepository contactEmailAddressRepository;
 
     @Override
     public ContactList getContactList(String jwt) {
@@ -47,11 +34,6 @@ public class ContactListServiceImpl implements ContactListService {
             return contactListRepository.getByUserId(user.getId());
         }
         return null;
-    }
-
-    @Override
-    public ContactList getContactList(ApplicationUser user) {
-        return contactListRepository.getByUserId(user.getId());
     }
 
     @Override
@@ -64,61 +46,6 @@ public class ContactListServiceImpl implements ContactListService {
         return null;
     }
 
-    @Override
-    public Contact addNewContact(String jwt, ContactDTO dto) {
-        ApplicationUser owner = userService.getUserByJwt(jwt);
-        if (owner != null) {
-            Contact contact = new Contact();
-            contact.setFirstName(dto.firstName);
-            contact.setLastName(dto.lastName);
-            contact.setContactList(getContactList(owner));
-            contact = contactRepository.save(contact);
-
-            Set<ContactEmailAddress> emailAddresses = new HashSet<>();
-            for (ContactEmailAddressDTO emailDTO : dto.getEmails()) {
-                ContactEmailAddress email = new ContactEmailAddress();
-                email.setContact(contact);
-                email.setEmail(emailDTO.email.toLowerCase());
-                email.setType(ContactTypeEnum.getEnumFromString(emailDTO.type));
-                email = contactEmailAddressRepository.save(email);
-                emailAddresses.add(email);
-            }
-
-            Set<ContactPhoneNumber> phoneNumbers = new HashSet<>();
-            for (ContactPhoneNumberDTO phoneNumberDTO : dto.getPhoneNumbers()) {
-                ContactPhoneNumber phoneNumber = new ContactPhoneNumber();
-                phoneNumber.setContact(contact);
-                phoneNumber.setPhoneNumber(phoneNumberDTO.phoneNumber);
-                phoneNumber.setType(ContactTypeEnum.getEnumFromString(phoneNumberDTO.type));
-                phoneNumber = contactPhoneNumberRepository.save(phoneNumber);
-                phoneNumbers.add(phoneNumber);
-            }
-
-            contact.setEmailAddresses(emailAddresses);
-            contact.setPhoneNumbers(phoneNumbers);
-            contact = contactRepository.save(contact);
-
-            checkForApplicationEmailMatch(contact.getId());
-            return contact;
-
-        } else {
-            return null;
-        }
-    }
-
-    private void checkForApplicationEmailMatch(Long contactId) {
-        Contact contact = contactRepository.get(contactId);
-        if (contact != null) {
-            for (ContactEmailAddress email : contact.getEmailAddresses()) {
-                ApplicationUser contactUser = userService.findByEmail(email.getEmail());
-                if (contactUser != null) {
-                    contact.setContactApplicationUser(contactUser);
-                    contactRepository.save(contact);
-                    break;
-                }
-            }
-        }
-    }
 
     @Override
     public Boolean deleteContact(String jwt, Long id) {
