+++ Backend/src/main/java/mygoogleserviceapi/contacts/service/interfaces/ContactListService.java
@@ -1,9 +1,7 @@
 package mygoogleserviceapi.contacts.service.interfaces;
 
-import mygoogleserviceapi.contacts.dto.ContactDTO;
 import mygoogleserviceapi.contacts.model.Contact;
 import mygoogleserviceapi.contacts.model.ContactList;
-import mygoogleserviceapi.shared.model.ApplicationUser;
 import org.springframework.core.io.Resource;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -13,11 +11,7 @@ public interface ContactListService {
 
     ContactList getContactList(String jwt);
 
-    ContactList getContactList(ApplicationUser user);
-
     List<Contact> getContacts(String jwt);
-
-    Contact addNewContact(String jwt, ContactDTO dto);
 
     Boolean deleteContact(String jwt, Long id);
