+++ Frontend/src/app/google-contacts/components/contact-person-page/contact-person-page.component.html
@@ -1,1 +1,2 @@
 <p>contact-person-page works!</p>
+<contact-profile-picture [id]="contact.id"></contact-profile-picture>
