+++ Frontend/src/app/google-contacts/components/contact-person-page/contact-person-page.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { Contact } from '../../model/contact';
+import { ContactService } from '../../services/contact.service';
 import { SearchContactsService } from '../contacts-header/search-contacts.service';
 
 @Component({
@@ -7,9 +10,21 @@ import { SearchContactsService } from '../contacts-header/search-contacts.servic
   styleUrls: ['./contact-person-page.component.scss'],
 })
 export class ContactPersonPageComponent implements OnInit {
-  constructor(private searchContactsService: SearchContactsService) {
+  contact: Contact = new Contact();
+
+  constructor(
+    private searchContactsService: SearchContactsService,
+    private route: ActivatedRoute,
+    private contactService: ContactService
+  ) {
     this.searchContactsService.announceSearchReset();
   }
 
-  ngOnInit(): void {}
+  ngOnInit(): void {
+    this.contact.id = this.route.snapshot.params['id'];
+    this.contactService.getContact(this.contact.id).subscribe((data) => {
+      this.contact = data;
+      console.log(this.contact);
+    });
+  }
 }
