+++ Frontend/src/app/google-contacts/services/contact-list.service.ts
@@ -18,4 +18,36 @@ export class ContactListService {
       headers: headers,
     });
   }
+
+  starContact(id: number) {
+    const url = this.url + '/star/' + id;
+    const headers = this.authService.getHeaders();
+    return this._http.put<any>(url, null, {
+      headers: headers,
+    });
+  }
+
+  unstarContact(id: number) {
+    const url = this.url + '/unstar/' + id;
+    const headers = this.authService.getHeaders();
+    return this._http.put<any>(url, null, {
+      headers: headers,
+    });
+  }
+
+  deleteContact(id: number) {
+    const url = this.url + '/delete/' + id;
+    const headers = this.authService.getHeaders();
+    return this._http.delete<any>(url, {
+      headers: headers,
+    });
+  }
+
+  deleteContactList(idList: number[]) {
+    const url = this.url + '/delete';
+    const headers = this.authService.getHeaders();
+    return this._http.put<any>(url, idList, {
+      headers: headers,
+    });
+  }
 }
