+++ Frontend/src/app/google-contacts/services/contact.service.ts
@@ -8,46 +8,22 @@ import { Contact } from '../model/contact';
   providedIn: 'root',
 })
 export class ContactService {
-  url = server + 'contacts';
+  url = server + 'contact';
 
   constructor(private _http: HttpClient, private authService: AuthService) {}
-
-  starContact(id: number) {
-    const url = this.url + '/star/' + id;
-    const headers = this.authService.getHeaders();
-    return this._http.put<any>(url, null, {
-      headers: headers,
-    });
-  }
-
-  unstarContact(id: number) {
-    const url = this.url + '/unstar/' + id;
-    const headers = this.authService.getHeaders();
-    return this._http.put<any>(url, null, {
-      headers: headers,
-    });
-  }
-
-  deleteContact(id: number) {
-    const url = this.url + '/delete/' + id;
-    const headers = this.authService.getHeaders();
-    return this._http.delete<any>(url, {
-      headers: headers,
-    });
-  }
-
-  deleteContactList(idList: number[]) {
-    const url = this.url + '/delete';
-    const headers = this.authService.getHeaders();
-    return this._http.put<any>(url, idList, {
-      headers: headers,
-    });
-  }
 
   createNewContact(contact: Contact) {
     const url = this.url + '/new';
     const headers = this.authService.getHeaders();
     return this._http.post<any>(url, contact, {
+      headers: headers,
+    });
+  }
+
+  getContact(id: number) {
+    const url = this.url + '/' + id;
+    const headers = this.authService.getHeaders();
+    return this._http.get<any>(url, {
       headers: headers,
     });
   }
